Move Configuration Menu to its own file
